refactor: consolidate tracker templates with existing builtin module
- Remove redundant src/plugins/trackers/builtin/templates.ts
- Update all trackers to import from src/templates/builtin.ts
- Add BEADS_RUST_TEMPLATE to the existing builtin templates
- Update template engine types and functions to support beads-rust
- Fix beads-bv test to match actual template content (no bd sync)
